feat(@angular-devkit/build-angular): code coverage option for ng build, ng serve and ng e2e
This commit allows --code-coverage and --code-coverage-exclude options
to be used with ng build, ng serve and ng e2e, so that it is easy to
instrument the code for coverage.

When used with these commands, the --code-coverage switch does not
automatically create a file with the report but allows the developers to
access the __coverage__ variable from Istanbul.js inside the browser to
get coverage data.
That object can then be used to generate a report as described in
https://istanbul.js.org/docs/advanced/coverage-object-report/
